pyproject.toml
[project]
name = "fo_services"
version = "0.1.0"
description = ""
authors = [
    { name = "Oliver Baumann", email = "oliver.baumann@uni-bayreuth.de" }
]
license = {text = "Proprietary"}
readme = "README.md"
package-mode = false
requires-python = ">=3.11"
dependencies = [
    "flask",
    "ldap3",
    "flask-restx",
    "flask-httpauth",
    "flask-login",
    "flask-sqlalchemy",
    "pandas",
    "uwsgi",
    "joblib",
    "matplotlib",
    "requests>=2.32.3",
    "psycopg[binary,pool]>=3.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
# testpaths = ['tests']
addopts = [
    "--import-mode=importlib",
    "--testdox"
]

[tool.coverage.run]
branch = true
source = ['fo_services']

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
    "coverage>=7.6.1",
    "pytest-mock>=3.14.0",
    "black>=24.8.0",
    "isort>=5.13.2",
    "pyflakes>=3.2.0",
    "ruff>=0.6.2",
    "pandas-stubs>=2.2.2.240807",
    "types-colorama>=0.4.15.20240311",
    "types-ldap3>=2.9.13.20240205",
    "types-requests>=2.32.0.20240712",
    "mypy>=1.11.2",
    "pytest-testdox>=3.1.0",
]

[tool.mypy]
disable_error_code = "import-untyped"

[tool.ruff]
indent-width = 4
fix = true
